+++ src/components/Icon/IconRegistry.ts
@@ -5,3 +5,6 @@ export const upArrow = `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 330
 export const menu = `<svg xmlns="http://www.w3.org/2000/svg" width="{{width}}" height="{{height}}" viewBox="0 -960 960 960" fill="{{color}}"><path d="M120-240v-80h720v80H120Zm0-200v-80h720v80H120Zm0-200v-80h720v80H120Z"/></svg>`;
 export const twitter = `<svg xmlns="http://www.w3.org/2000/svg"  viewBox="0 0 30 30" width="{{width}}" height="{{height}}" fill="{{color}}"><path d="M28,6.937c-0.957,0.425-1.985,0.711-3.064,0.84c1.102-0.66,1.947-1.705,2.345-2.951c-1.03,0.611-2.172,1.055-3.388,1.295 c-0.973-1.037-2.359-1.685-3.893-1.685c-2.946,0-5.334,2.389-5.334,5.334c0,0.418,0.048,0.826,0.138,1.215 c-4.433-0.222-8.363-2.346-10.995-5.574C3.351,6.199,3.088,7.115,3.088,8.094c0,1.85,0.941,3.483,2.372,4.439 c-0.874-0.028-1.697-0.268-2.416-0.667c0,0.023,0,0.044,0,0.067c0,2.585,1.838,4.741,4.279,5.23 c-0.447,0.122-0.919,0.187-1.406,0.187c-0.343,0-0.678-0.034-1.003-0.095c0.679,2.119,2.649,3.662,4.983,3.705 c-1.825,1.431-4.125,2.284-6.625,2.284c-0.43,0-0.855-0.025-1.273-0.075c2.361,1.513,5.164,2.396,8.177,2.396 c9.812,0,15.176-8.128,15.176-15.177c0-0.231-0.005-0.461-0.015-0.69C26.38,8.945,27.285,8.006,28,6.937z"/></svg>`;
 export const youtube = `<svg width="{{width}}" height="{{height}}" fill="{{color}}" version="1.1" id="Layer_1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" viewBox="0 0 461.001 461.001" xml:space="preserve"><g><path d="M365.257,67.393H95.744C42.866,67.393,0,110.259,0,163.137v134.728 c0,52.878,42.866,95.744,95.744,95.744h269.513c52.878,0,95.744-42.866,95.744-95.744V163.137 C461.001,110.259,418.135,67.393,365.257,67.393z M300.506,237.056l-126.06,60.123c-3.359,1.602-7.239-0.847-7.239-4.568V168.607 c0-3.774,3.982-6.22,7.348-4.514l126.06,63.881C304.363,229.873,304.298,235.248,300.506,237.056z"/></g></svg>`;
+export const logo = `
+    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 150.12 27.31" width="{{width}}" height="{{height}}" fill="{{color}}"><defs><style>.cls-1,.cls-2{fill:#333132;}.cls-1,.cls-3{fill-rule:evenodd;}.cls-3{fill:#523181;}</style></defs><g id="Layer_1" data-name="Layer 1"><path class="cls-1" d="M8,22.08H2.37a1.65,1.65,0,0,1-1.43-.78L.72,21a2.15,2.15,0,0,1-.18-1.61A2.53,2.53,0,0,1,.72,19c.32-.67.63-1.35,1-2s.59-1.12.87-1.69.38-.86.59-1.29.55-1.11.83-1.66c.07-.15.12-.3.19-.44.25-.51.51-1,.75-1.55A1,1,0,0,0,5,9.94C5,8.64,5,7.34,5,6a1.12,1.12,0,0,1,.22-.81A1.1,1.1,0,0,1,5.48,5a2.87,2.87,0,0,1,1.17-.17c.94,0,1.88,0,2.83,0a2.38,2.38,0,0,1,.88.18,1,1,0,0,1,.57,1c0,1.27,0,2.55,0,3.81a2.57,2.57,0,0,0,.44,1.38c.13.23.23.47.34.71l.26.5c.21.44.43.87.64,1.3l.78,1.69c.21.43.43.84.64,1.26.08.18.16.36.25.53.31.64.64,1.27.91,1.92a3,3,0,0,1,.18,1,2.17,2.17,0,0,1-.68,1.51,1.32,1.32,0,0,1-1,.42H8Zm-.1-6.2h3.53c.29,0,.39-.15.26-.42l-.57-1.13c-.08-.17-.13-.36-.21-.54-.3-.68-.62-1.36-.92-2-.1-.22-.21-.43-.3-.65a2.27,2.27,0,0,1-.24-.78c0-1.26,0-2.52,0-3.78,0-.4-.08-.49-.48-.48H6.85c-.38,0-.43,0-.43.41v3.3a2.26,2.26,0,0,1-.17,1.12c-.24.46-.4.94-.62,1.4-.29.66-.6,1.31-.9,2-.19.41-.39.82-.57,1.23a.38.38,0,0,0,0,.29c0,.06.17.1.26.1H7.9"/><path class="cls-2" d="M79.37,19.48A7,7,0,0,1,77,21.22,9.25,9.25,0,0,1,73,22a6.67,6.67,0,0,1-4.91-1.86,6.48,6.48,0,0,1-1.89-4.83V11.87A6.58,6.58,0,0,1,68,7.05a6.37,6.37,0,0,1,4.75-1.87A6.69,6.69,0,0,1,77.5,6.64a5.88,5.88,0,0,1,1.78,4l0,.06H75.7a3.35,3.35,0,0,0-.83-1.94,2.69,2.69,0,0,0-2-.65,2.56,2.56,0,0,0-2.13,1.05A4.3,4.3,0,0,0,70,11.85v3.43a4.39,4.39,0,0,0,.8,2.74,2.65,2.65,0,0,0,2.23,1,6.37,6.37,0,0,0,1.64-.17,2.57,2.57,0,0,0,.94-.45V15.86H73V13.35h6.37Z"/><path class="cls-2" d="M95.18,15a6.83,6.83,0,0,1-1.92,5,6.66,6.66,0,0,1-5,1.95,6.65,6.65,0,0,1-7-6.93V12.11a6.81,6.81,0,0,1,1.92-5,7.38,7.38,0,0,1,10,0,6.82,6.82,0,0,1,1.93,5Zm-3.78-3a4.74,4.74,0,0,0-.83-2.87,3,3,0,0,0-4.67,0,4.76,4.76,0,0,0-.81,2.87v3a4.79,4.79,0,0,0,.82,2.9,2.78,2.78,0,0,0,2.36,1.12,2.72,2.72,0,0,0,2.31-1.12A4.79,4.79,0,0,0,91.4,15Z"/><polygon class="cls-2" points="102.99 17.31 103.1 17.9 103.17 17.9 103.28 17.35 106.52 5.41 110.49 5.41 105.2 21.74 101.11 21.74 95.84 5.41 99.79 5.41 102.99 17.31"/><polygon class="cls-2" points="113.96 20.59 122.03 20.59 122.03 21.74 112.62 21.74 112.62 5.41 113.96 5.41 113.96 20.59"/><path class="cls-2" d="M133.69,17.18h-7.47l-1.69,4.56h-1.39l6.2-16.33h1.27l6.15,16.33h-1.39Zm-7-1.2h6.6L130,7.15h-.06Z"/><path class="cls-2" d="M139.26,21.74V5.41H144a6.49,6.49,0,0,1,4,1.07,3.79,3.79,0,0,1,1.42,3.23,3.53,3.53,0,0,1-.71,2.19,3.68,3.68,0,0,1-1.94,1.29,3.89,3.89,0,0,1,2.42,1.4,4,4,0,0,1,.94,2.61,4.15,4.15,0,0,1-1.42,3.36,5.79,5.79,0,0,1-3.8,1.18Zm1.34-9h3.75a4.26,4.26,0,0,0,2.74-.78,2.74,2.74,0,0,0,1-2.29,2.71,2.71,0,0,0-1-2.32,4.92,4.92,0,0,0-3-.78H140.6Zm0,1.15v6.71h4.3a4.17,4.17,0,0,0,2.84-.89,3.09,3.09,0,0,0,1-2.48,3.22,3.22,0,0,0-.94-2.4,3.64,3.64,0,0,0-2.71-.94Z"/><polygon class="cls-2" points="35.01 6.52 29.37 6.52 29.37 21.79 27.99 21.79 27.99 6.52 22.36 6.52 22.36 5.35 35.01 5.35 35.01 6.52"/><polygon class="cls-2" points="49.85 21.79 48.45 21.79 48.45 13.91 39.05 13.91 39.05 21.79 37.66 21.79 37.66 5.35 39.05 5.35 39.05 12.73 48.45 12.73 48.45 5.35 49.85 5.35 49.85 21.79"/><polygon class="cls-2" points="63.11 13.91 55.42 13.91 55.42 20.62 64.26 20.62 64.26 21.79 54.03 21.79 54.03 5.35 64.2 5.35 64.2 6.52 55.42 6.52 55.42 12.73 63.11 12.73 63.11 13.91"/><path class="cls-3" d="M5.89,24.23h8.45a1.56,1.56,0,0,1,.94.18A1.59,1.59,0,0,1,16,25.83a1.55,1.55,0,0,1-1.58,1.48H1.7a1.53,1.53,0,0,1-1.32-.57,1.45,1.45,0,0,1,0-2,1.37,1.37,0,0,1,1.14-.54H5.89"/><path class="cls-3" d="M1.54,0A1.37,1.37,0,0,0,.4.54a1.46,1.46,0,0,0,0,2,1.55,1.55,0,0,0,1.32.56H14.46A1.56,1.56,0,0,0,15.28.2a1.58,1.58,0,0,0-.94-.2Z"/></g></svg>
+`;
